Basic Host/Guest controllers, all implemented methods tested.
Also some changes to the DAOs, which are tested second hand here: added a read() method that does a lookup based on username and password, which required the checkpassword() method to be changed to return a dummy true value until passwords can be properly implemented. Also changed the return types on the delete() methods from void to boolean so that I can indicate whether the deletion took or not.