Genericized users
Created a new abstract class User that now serves as the parent class for Guest and Host. I was able to clean up a lot of code as a result (though I also had to suppress a lot of warnings; looking for fixes on that). Also added logout functionality to the dashboards.