Attempting a one page app
When I started to add bootstrap, it became apparent that it might just be easier to switch to using a one page app. This would also likely make it easier to add AJAX into the project in more than just a passing, "I'm doing this just to meet requirements" sort of way. Kicking this onto it's own branch so I can dispose of it in case this all goes terribly sideways.